+++ backend/src/db/strategies/__init__.py
@@ -2,7 +2,7 @@
 from abc import ABC, abstractmethod
 # from io import BytesIO
 
-from src.db import PineconeDB, RedisDB
+from src.db import PineconeDB, RedisDB, PGVectorDB
 
 
 # Define the strategy interface
@@ -13,6 +13,10 @@ class VectorStoreStrategy(ABC):
 
     @abstractmethod
     def load(self):
+        pass
+    
+    @abstractmethod
+    def delete(self):
         pass
 
 #########################################################################
@@ -75,9 +79,40 @@ class PineconeStrategy(VectorStoreStrategy):
             embeddings=self.embeddings,
             namespace=self.namespace
         )
+        
+    def delete(self):
+        return self.service.delete()
+        
+#########################################################################
+## Postgres Strategy
+#########################################################################
+class PostgresStrategy(VectorStoreStrategy):
+    def __init__(
+        self,
+        connection: str,
+        collection_name: str,
+        embeddings = None,
+    ):
+        self.connection = connection
+        self.collection_name = collection_name
+        self.embeddings = embeddings,
+        self.service = PGVectorDB(
+			connection=self.connection,
+			collection_name=self.collection_name,
+            embeddings=self.embeddings
+		)
+
+    def add(self, documents):
+        return self.service.add_docs(documents)
+
+    def load(self):
+        return self.service.from_existing()
+    
+    def delete(self):
+        return self.service.delete()
 
 #########################################################################
-## Pinecone Strategy
+## Redis Strategy
 #########################################################################
 class RedisStrategy(VectorStoreStrategy):
     def __init__(
@@ -107,6 +142,9 @@ class RedisStrategy(VectorStoreStrategy):
         return self.service.from_existing(
             schema=self.index_schema
         )
+        
+    def delete(self):
+        return self.service.delete(self.index_name)
 
 #########################################################################
 ## Strategy Context
@@ -126,3 +164,6 @@ class VectorstoreContext:
 
     def load(self):
         return self.strategy.load()
+
+    def delete(self):
+        return self.strategy.delete()